refactor: move pomodoro code
